requirements.txt
# Wi-Fi CSI 기반 비접촉 실내 이상상황 감지 시스템
# 의존성 목록

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# 신호처리
scikit-learn>=1.3.0

# 딥러닝
tensorflow>=2.13.0

# 시리얼 통신 (ESP32)
pyserial>=3.5

# 웹 대시보드
flask>=3.0.0
flask-socketio>=5.3.0

# 시각화
matplotlib>=3.7.0
seaborn>=0.12.0
